+++ task_yell/src/components/priority-colors.ts
@@ -1,0 +1,7 @@
+import { Priority } from "@/lib/types";
+
+export const priorityColors: Record<Priority, string> = {
+  low: "bg-blue-100 dark:bg-blue-900",
+  medium: "bg-yellow-100 dark:bg-yellow-900",
+  high: "bg-red-100 dark:bg-red-900",
+};
